Finished the backend code

+++ Backend/routes/visits.js
@@ -2,21 +2,27 @@ const express = require("express");
 const router = express.Router();
 const Visits = require("../models/visits/Visits");
 
-// Adds a patient
+// Adds a Visit
 router.post("/api/visits/", async (req, res) => {
-  const visit = new Visits({
-    patientId: req.body.patientId,
-    patientName: req.body.patientName,
-    patientSymptoms: req.body.patientSymptoms,
-    patientDiagnosis: req.body.patientDiagnosis,
-    patientVisitDate: req.body.patientVisitDate,
-    patientMedications: req.body.patientMedications
-    // patientNextVisitDate: req.body.patientNextVisitDate
-  });
   try {
-    const savedVisit = await visit.save();
-    res.json(savedVisit);
-    console.log(savedVisit._id);
+    if (req.user) {
+      const visit = new Visits({
+        doctorId: req.user._id,
+        patientId: req.body.patientId,
+        patientName: req.body.patientName,
+        patientSymptoms: req.body.patientSymptoms,
+        patientDiagnosis: req.body.patientDiagnosis,
+        patientVisitDate: req.body.patientVisitDate,
+        patientMedications: req.body.patientMedications,
+        visitCost: req.body.visitCost,
+      });
+
+      const savedVisit = await visit.save();
+      res.json(savedVisit);
+      console.log(savedVisit._id);
+    } else {
+      return res.status(401).json({ message: 'Invalid token' });
+    }
   } catch (err) {
     res.json({
       msg: err
@@ -27,8 +33,12 @@ router.post("/api/visits/", async (req, res) => {
 // Gets all visits
 router.get("/api/visits/", async (req, res) => {
   try {
-    const visits = await Visits.find();
-    res.json(visits);
+    if (req.user) {
+      const visits = await Visits.find({ doctorId: req.user._id });
+      res.json(visits);
+    } else {
+      return res.status(401).json({ message: 'Invalid token' });
+    }
   } catch (err) {
     res.json({
       msg: err
@@ -39,8 +49,15 @@ router.get("/api/visits/", async (req, res) => {
 // Get a specific visit
 router.get("/api/visits/:visitId", async (req, res) => {
   try {
-    const visit = await Visits.findById(req.params.visitId);
-    res.json(visit);
+    if (req.user) {
+      const visit = await Visits.find({
+        _id: req.params.visitId,
+        doctorId: req.user._id
+      });
+      res.json(visit);
+    } else {
+      return res.status(401).json({ message: 'Invalid token' });
+    }
   } catch (err) {
     res.json({
       msg: err
@@ -48,39 +65,41 @@ router.get("/api/visits/:visitId", async (req, res) => {
   };
 })
 
-// Get visits from a specific visit
-router.get("/api/visits/patients/:patientId", async (req,res) => {
-  try {
-    const visit = await Visits.find({ patientId: req.params.patientId })
-    res.json(visit)
-  } catch (err) {
-    res.json({
-      msg: err
-    })
-  }
-})
-
-// Update a specific patient
+// Update a specific Visit
 router.patch("/api/visits/:visitId", async (req, res) => {
   try {
-    const updatedVisit = await Visits.findById(req.params.visitId);
-    updatedVisit.patientSymptoms = req.body.patientSymptoms
-    updatedVisit.patientDiagnosis = req.body.patientDiagnosis
-    updatedVisit.patientMedications = req.body.patientMedications
-    await updatedVisit.save();
-    res.json(updatedVisit);
+    if (req.user) {
+      const updatedVisit = await Visits.find({
+        _id: req.params.visitId,
+        doctorId: req.user._id
+      });
+      updatedVisit.patientSymptoms = req.body.patientSymptoms
+      updatedVisit.patientDiagnosis = req.body.patientDiagnosis
+      updatedVisit.patientMedications = req.body.patientMedications
+      updatedVisit.visitCost = req.body.visitCost
+
+      await updatedVisit.save();
+      res.json(updatedVisit);
+    } else {
+      return res.status(401).json({ message: 'Invalid token' });
+    }
   } catch (err) {
     res.json({ msg: err });
   }
 });
 
-// Delete a specific patient
+// Delete a specific Visit
 router.delete("/api/visits/:visitId", async (req, res) => {
   try {
-    const removedVisit = await Visits.deleteOne({
-      _id: req.params.visitId
-    });
-    res.json(removedVisit);
+    if (req.user) {
+      const removedVisit = await Visits.deleteOne({
+        _id: req.params.visitId,
+        doctorId: req.user._id
+      });
+      res.json(removedVisit);
+    } else {
+      return res.status(401).json({ message: 'Invalid token' });
+    }
   } catch (err) {
     res.json({ msg: err });
   }
